Windows port: transformer: fix MSVC narrowing and GNU-builtin errors
- sdpa/device/exp_ring_joint_sdpa_program_factory.cpp: static_cast size_t
  ring_size/logical_n/q_chunk_start values to uint32_t in braced kernel
  compile-time-arg vectors and struct initializers (C2397/C2398)
- sdpa/device/ring_joint_sdpa_device_operation.cpp: static_cast size_t
  num_devices to uint32_t when building RingAttentionAllGatherAsyncParams
  (C2398)
- sdpa/device/ring_joint_sdpa_program_factory.cpp: static_cast size_t
  grid_size.x arithmetic to uint32_t for num_receivers (C2397)
- sdpa_decode/device/sdpa_decode_device_operation.hpp,
  sdpa_decode_program_factory.cpp: replace __builtin_clz/__builtin_ctz
  with the portable tt::compiler::count_leading_zeros32/
  count_trailing_zeros32 helpers from hostdevcommon/tt_compiler.h
  (C3861, plus the C2737 this masked)

+++ ttnn/cpp/ttnn/operations/transformer/sdpa/device/exp_ring_joint_sdpa_program_factory.cpp
@@ -526,7 +526,7 @@ tt::tt_metal::ProgramDescriptor build_exp_ring_joint_sdpa_program_descriptor(
         num_local_k_chunks,
         num_joint_k_chunks,
         num_q_chunks,
-        args.ring_size,
+        static_cast<uint32_t>(args.ring_size),
         qk_out_subblock_h};
 
     TensorAccessorArgs(input_tensor_q.buffer()).append_to(reader_compile_time_args);
@@ -625,7 +625,7 @@ tt::tt_metal::ProgramDescriptor build_exp_ring_joint_sdpa_program_descriptor(
         Sk_chunk_t,
         local_padded_N,
         local_padded_Nt,
-        args.logical_n,
+        static_cast<uint32_t>(args.logical_n),
         logical_nt,
         Lt,
         L,
@@ -636,7 +636,7 @@ tt::tt_metal::ProgramDescriptor build_exp_ring_joint_sdpa_program_descriptor(
         num_q_chunks,
         packed_identity_scalar,
         scale_packed,
-        args.ring_size,
+        static_cast<uint32_t>(args.ring_size),
         global_n_partial_col,
         joint_l_partial_col,
         static_cast<std::uint32_t>(use_streaming_compute),
@@ -655,13 +655,13 @@ tt::tt_metal::ProgramDescriptor build_exp_ring_joint_sdpa_program_descriptor(
         Sk_chunk_t,
         local_padded_N,
         local_padded_Nt,
-        args.logical_n,
+        static_cast<uint32_t>(args.logical_n),
         logical_nt,
         Lt,
         L,
         num_local_k_chunks,
         num_joint_k_chunks,
-        args.ring_size,
+        static_cast<uint32_t>(args.ring_size),
         qk_out_subblock_w,
         qk_out_subblock_h,
         out_out_subblock_w,
@@ -1117,7 +1117,7 @@ tt::tt_metal::ProgramDescriptor build_exp_ring_joint_sdpa_program_descriptor(
             work.head_work.push_back(CoreHeadWork{
                 .batch = head_id / NH,
                 .head = head_id % NH,
-                .q_chunk_start = (seg_id % segs_per_head) * sdpa_grid.x + core.x,
+                .q_chunk_start = static_cast<uint32_t>((seg_id % segs_per_head) * sdpa_grid.x + core.x),
                 .q_chunk_count = 1,
             });
         }
